backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
aiosqlite==0.20.0
authlib==1.3.2
httpx==0.27.2
python-jose[cryptography]==3.3.0
